particular login user can create here own data and delete.

+++ backend/controllers/expenseControllers.js
@@ -9,6 +9,7 @@ exports.addExpense = async (req, res) => {
     category,
     description,
     date,
+    user: req.user._id,
   });
 
   try {
@@ -16,27 +17,27 @@ exports.addExpense = async (req, res) => {
     if (!title || !category || !description || !date) {
       return res.status(400).json({ message: "All fields are required!" });
     }
-    if (amount <= 0 ||  !amount == "number") {
+    if (amount <= 0 || !amount == "number") {
       return res
         .status(400)
         .json({ message: "Amount must be a positive number!" });
     }
 
     await expense.save();
-    res.status(200).json({ success: true, message: "Expense Added" ,expense});
+    res.status(200).json({ success: true, message: "Expense Added", expense });
   } catch (error) {
     res.status(500).json({ message: "Server Error" });
   }
-
-  console.log(expense);
 };
 
 exports.getExpense = async (req, res) => {
   try {
-    const expenses = await ExpenseSchema.find().sort({ createdAt: -1 });
-    if (!expenses.length) {
-        return res.status(404).json({ message: "No expenses found" });
-      }
+    const expenses = await ExpenseSchema.find({ user: req.user._id }).sort({
+      createdAt: -1,
+    });
+    if (!expenses || expenses.length === 0) {
+      return res.status(404).json({ message: "No expenses found" });
+    }
     res.status(200).json(expenses);
   } catch (error) {
     res.status(500).json({ message: "Server Error" });
@@ -46,11 +47,21 @@ exports.getExpense = async (req, res) => {
 exports.deleteExpense = async (req, res) => {
   const { id } = req.params;
   try {
-    const expense = await ExpenseSchema.findByIdAndDelete(id);
+    const expense = await ExpenseSchema.findById(id);
     if (!expense) {
       return res.status(404).json({ message: "Expense not found" });
     }
-    res.status(200).json({success: true, message: "Expense Deleted" });
+
+    // Check if the user deleting the income is the same as the one who created it
+    if (expense.user.toString() !== req.user._id.toString()) {
+      return res
+        .status(403)
+        .json({ message: "Not authorized to delete this expense" });
+    }
+
+    await ExpenseSchema.findByIdAndDelete(id);
+
+    res.status(200).json({ success: true, message: "Expense Deleted" });
   } catch (err) {
     res.status(500).json({ message: "Server Error" });
   }
